requirements.txt
httpx
mcp
python-dotenv
pytest
